Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
gtts
pypdf
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
google-generativeai
edge-tts
pdf2image
Pillow

